refactor(faq): split admin tasks
